requirements.txt
python-telegram-bot==22.7
python-dotenv==1.0.0
APScheduler==3.10.4
pytz==2024.1
aiosqlite==0.19.0
